more cleanup

+++ src/it/itmain.c
@@ -591,7 +591,8 @@ sb32 itMainMakeContainerItem(GObj *parent_gobj)
         if (item_kind <= nITKindCommonEnd)
         {
             vel.x = 0.0F;
-            vel.y = *(f32*) ((intptr_t)&lITMainContainerVelocitiesY + ((uintptr_t) &((u32*)gITManagerFileData)[item_kind])); // Linker thing; quite ridiculous especially since lITMainContainerVelocitiesY is 0
+            // Quite ridiculous especially since lITMainContainerVelocitiesY is 0
+            vel.y = *(f32*) ((intptr_t)&lITMainContainerVelocitiesY + ((uintptr_t) &((f32*)gITManagerFileData)[item_kind]));
             vel.z = 0;
 
             if (itManagerMakeItemSetupCommon(parent_gobj, item_kind, &DObjGetStruct(parent_gobj)->translate.vec.f, &vel, (ITEM_FLAG_COLLPROJECT | ITEM_FLAG_PARENT_ITEM)) != NULL)
